--- src/main/java/roomescape/AutoIncrementId.java
@@ -1,16 +1,0 @@
-package roomescape;
-
-import java.util.concurrent.atomic.AtomicLong;
-
-public class AutoIncrementId {
-    private final AtomicLong nextValue;
-
-    public AutoIncrementId() {
-        this.nextValue = new AtomicLong(1);
-    }
-
-    public Long getNextId(){
-        return nextValue.getAndIncrement();
-    }
-
-}
