+++ src/main/java/roomescape/controller/GlobalExceptionHandler.java
@@ -1,0 +1,21 @@
+package roomescape.controller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import roomescape.NotFoundReservationException;
+
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+
+    @ExceptionHandler(IllegalArgumentException.class)
+    public ResponseEntity handleIllegalArgumentException(IllegalArgumentException e) {
+        return ResponseEntity.badRequest().body(e.getMessage());
+    }
+
+    @ExceptionHandler(NotFoundReservationException.class)
+    public ResponseEntity handleException(NotFoundReservationException e) {
+        return ResponseEntity.badRequest().build();
+    }
+
+}
