+++ src/main/java/roomescape/controller/RoomscapeController.java
@@ -1,0 +1,26 @@
+package roomescape.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/")
+public class RoomscapeController {
+
+    @GetMapping
+    public String home() {
+        return "home";
+    }
+
+    @GetMapping("/reservation")
+    public String admin() {
+        return "reservation";
+    }
+
+    @GetMapping("/time")
+    public String time() {
+        return "time";
+    }
+
+}
