+++ src/main/java/roomescape/dao/ReservationDao.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.dao;
 
 import java.util.List;
 import java.util.Map;
@@ -7,6 +7,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
+import roomescape.domain.Reservation;
 
 @Repository
 public class ReservationDao {
