+++ src/test/java/roomescape/ReservationDaoTest.java
@@ -1,13 +1,12 @@
 package roomescape;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.test.context.SpringBootTest;
+import roomescape.dao.ReservationDao;
+import roomescape.domain.Reservation;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
 class ReservationDaoTest {
